+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* Stored CELO base value in env variables
 * Fixed copy method to copy each address seperately
 * Fixed collapse to open every collapse element individually
 * Added query to display Validator Votes List data 
@@ -12,6 +13,7 @@
 * [#68] Added transaction contract input parameters in TransactionDetails
 * Hidden Validator Votes
 * [#74] Changed token name from cGLD to CELO 
+
 
 ## v0.1.2