+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 
 * Stored CELO base value in env variables
+* Fixed copy method to copy each address seperately
+* Fixed collapse to open every collapse element individually
+* Added query to display Validator Votes List data 
 
 ## v0.1.3