+++ CHANGELOG.md
@@ -6,10 +6,18 @@
 
 ## v0.1.3
 
+* Stored CELO base value in env variables
+* Fixed copy method to copy each address seperately
+* Fixed collapse to open every collapse element individually
+* Added query to display Validator Votes List data 
+
+## v0.1.3
+
 * [#69] Hidden Search Bar and Network Selection
 * [#68] Added transaction contract input parameters in TransactionDetails
 * Hidden Validator Votes
 * [#74] Changed token name from cGLD to CELO 
+
 
 ## v0.1.2